fix(win): 修复窗口恢复代码 DWORD→LONG 收窄转换导致 Windows 构建失败
RECT 列表初始化的 left/top/sum 为 DWORD/unsigned long，赋给 LONG 字段触发
C4838(收窄转换)，MSVC 按 C2220 视为错误中断出包。改为对全部操作数显式
static_cast<LONG>，保持恢复矩形语义不变。

+++ windows/runner/win32_window.cpp
@@ -120,8 +120,10 @@ void RestoreWindowFrame(HWND hwnd) {
   if (!ok || width == 0 || height == 0) {
     return;
   }
-  RECT rect{left, top, left + static_cast<LONG>(width),
-            top + static_cast<LONG>(height)};
+  RECT rect{static_cast<LONG>(left),
+            static_cast<LONG>(top),
+            static_cast<LONG>(left) + static_cast<LONG>(width),
+            static_cast<LONG>(top) + static_cast<LONG>(height)};
   if (MonitorFromRect(&rect, MONITOR_DEFAULTTONULL) == nullptr) {
     return;
   }
